Moved cli serviceinit functions from marathon_serviceinit to status.py
