+++ paasta_tools/cli/cmds/status.py
@@ -56,11 +56,6 @@ from paasta_tools.flink_tools import FlinkDeploymentConfig
 from paasta_tools.kafkacluster_tools import KafkaClusterDeploymentConfig
 from paasta_tools.kubernetes_tools import KubernetesDeploymentConfig
 from paasta_tools.kubernetes_tools import KubernetesDeployStatus
-from paasta_tools.marathon_serviceinit import bouncing_status_human
-from paasta_tools.marathon_serviceinit import desired_state_human
-from paasta_tools.marathon_serviceinit import haproxy_backend_report
-from paasta_tools.marathon_serviceinit import marathon_app_deploy_status_human
-from paasta_tools.marathon_serviceinit import status_marathon_job_human
 from paasta_tools.marathon_tools import MarathonDeployStatus
 from paasta_tools.marathon_tools import MarathonServiceConfig
 from paasta_tools.mesos_tools import format_tail_lines_for_mesos_task
@@ -73,6 +68,7 @@ from paasta_tools.utils import DEFAULT_SOA_DIR
 from paasta_tools.utils import format_table
 from paasta_tools.utils import get_soa_cluster_deploy_files
 from paasta_tools.utils import InstanceConfig
+from paasta_tools.utils import is_under_replicated
 from paasta_tools.utils import list_all_instances_for_service
 from paasta_tools.utils import list_clusters
 from paasta_tools.utils import list_services
@@ -1339,3 +1335,108 @@ def paasta_status(args) -> int:
             return_codes.append(return_code)
 
     return max(return_codes)
+
+
+def bouncing_status_human(app_count, bounce_method):
+    if app_count == 0:
+        return PaastaColors.red("Disabled")
+    elif app_count == 1:
+        return PaastaColors.green("Configured")
+    elif app_count > 1:
+        return PaastaColors.yellow("Bouncing (%s)" % bounce_method)
+    else:
+        return PaastaColors.red("Unknown (count: %s)" % app_count)
+
+
+def desired_state_human(desired_state, instances):
+    if desired_state == "start" and instances != 0:
+        return PaastaColors.bold("Started")
+    elif desired_state == "start" and instances == 0:
+        return PaastaColors.bold("Stopped")
+    elif desired_state == "stop":
+        return PaastaColors.red("Stopped")
+    else:
+        return PaastaColors.red("Unknown (desired_state: %s)" % desired_state)
+
+
+def haproxy_backend_report(normal_instance_count, up_backends):
+    """Given that a service is in smartstack, this returns a human readable
+    report of the up backends"""
+    # TODO: Take into account a configurable threshold, PAASTA-1102
+    crit_threshold = 50
+    under_replicated, ratio = is_under_replicated(
+        num_available=up_backends,
+        expected_count=normal_instance_count,
+        crit_threshold=crit_threshold,
+    )
+    if under_replicated:
+        status = PaastaColors.red("Critical")
+        count = PaastaColors.red(
+            "(%d/%d, %d%%)" % (up_backends, normal_instance_count, ratio)
+        )
+    else:
+        status = PaastaColors.green("Healthy")
+        count = PaastaColors.green("(%d/%d)" % (up_backends, normal_instance_count))
+    up_string = PaastaColors.bold("UP")
+    return f"{status} - in haproxy with {count} total backends {up_string} in this namespace."
+
+
+def marathon_app_deploy_status_human(status, backoff_seconds=None):
+    status_string = MarathonDeployStatus.tostring(status)
+
+    if status == MarathonDeployStatus.Waiting:
+        deploy_status = (
+            "%s (new tasks waiting for capacity to become available)"
+            % PaastaColors.red(status_string)
+        )
+    elif status == MarathonDeployStatus.Delayed:
+        deploy_status = "{} (tasks are crashing, next won't launch for another {} seconds)".format(
+            PaastaColors.red(status_string), backoff_seconds
+        )
+    elif status == MarathonDeployStatus.Deploying:
+        deploy_status = PaastaColors.yellow(status_string)
+    elif status == MarathonDeployStatus.Stopped:
+        deploy_status = PaastaColors.grey(status_string)
+    elif status == MarathonDeployStatus.Running:
+        deploy_status = PaastaColors.bold(status_string)
+    else:
+        deploy_status = status_string
+
+    return deploy_status
+
+
+def status_marathon_job_human(
+    service: str,
+    instance: str,
+    deploy_status: str,
+    desired_app_id: str,
+    app_count: int,
+    running_instances: int,
+    normal_instance_count: int,
+) -> str:
+    name = PaastaColors.cyan(compose_job_id(service, instance))
+
+    if app_count >= 0:
+        if running_instances >= normal_instance_count:
+            status = PaastaColors.green("Healthy")
+            instance_count = PaastaColors.green(
+                "(%d/%d)" % (running_instances, normal_instance_count)
+            )
+        elif running_instances == 0:
+            status = PaastaColors.yellow("Critical")
+            instance_count = PaastaColors.red(
+                "(%d/%d)" % (running_instances, normal_instance_count)
+            )
+        else:
+            status = PaastaColors.yellow("Warning")
+            instance_count = PaastaColors.yellow(
+                "(%d/%d)" % (running_instances, normal_instance_count)
+            )
+        return "Marathon:   {} - up with {} instances. Status: {}".format(
+            status, instance_count, deploy_status
+        )
+    else:
+        status = PaastaColors.yellow("Warning")
+        return "Marathon:   {} - {} (app {}) is not configured in Marathon yet (waiting for bounce)".format(
+            status, name, desired_app_id
+        )
